refactor(taiko-client-rs): consolidate whitelist preconfirmation driver helpers
Final consolidation pass of the whitelist driver simplification plan
(follow-up to #21768 and #21769):

- Extract the duplicated TaikoPayloadAttributes construction (importer
  envelope path vs API build path) into one shared payload::build_driver_payload;
  the API service reuses its validation stub payload for both call sites.
- Merge importer/response.rs into importer/ingress.rs and unify the
  request-topic and EOS-request serving flows into one
  serve_envelope_by_hash (recent cache -> L2 rebuild -> publish).
- Remove the request-hash dedupe tracker (WindowedHashTracker): the
  message-id function hashes topic + data, so gossipsub's 120s duplicate
  cache already drops identical requests before they reach the handler;
  per-peer token-bucket rate limiting remains.
- Drop redundant metric families: cache_import_attempts_total (sum of
  cache_import_results_total) and rpc_requests_total (the rpc_duration
  histogram _count); fold driver_submit_total into a result-labelled
  driver_submit_duration_seconds histogram.
- Merge the three REST/WS startup error variants into one
  RestWsServerStartup(String).
- Fold api/interface.rs into api/mod.rs, state.rs into server/mod.rs,
  and http_utils.rs + http_error.rs into one http.rs; merge the twin
  inbound gossip log functions; inline the sync_ready_transition and
  confirmed_boundary_or_genesis one-liners.

No wire-format, REST-contract, or behavior changes.
