+++ packages/taiko-client-rs/crates/whitelist-preconfirmation-driver/src/api/server/handlers.rs
@@ -8,10 +8,8 @@ use axum::{
 };
 
 use super::{
-    PRECONF_BLOCKS_BODY_LIMIT_BYTES,
-    http_error::ApiHttpError,
-    http_utils::{error_response, json_response, no_content_response, read_request_body},
-    state::AppState,
+    AppState, PRECONF_BLOCKS_BODY_LIMIT_BYTES,
+    http::{ApiHttpError, error_response, json_response, no_content_response, read_request_body},
     websocket::serve_websocket_notifications,
 };
 use crate::{api::types::BuildPreconfBlockRequest, metrics::WhitelistPreconfirmationDriverMetrics};
