+++ packages/taiko-client-rs/crates/whitelist-preconfirmation-driver/src/importer/ingress.rs
@@ -1,4 +1,4 @@
-//! Ingress entrypoints for unsafe payload and response handling.
+//! Ingress entrypoints for unsafe payload handling and request/response serving.
 //!
 //! Signature authenticity and operator-set membership are enforced at gossip
 //! acceptance time in the network layer's inbound validation
@@ -7,14 +7,22 @@
 
 use std::sync::Arc;
 
-use alloy_primitives::B256;
+use alloy_consensus::TxEnvelope;
+use alloy_eips::Encodable2718;
+use alloy_primitives::{B256, Bytes, U256};
 use alloy_provider::Provider;
-use tracing::debug;
+use alloy_rpc_types::eth::Transaction as RpcTransaction;
+use protocol::codec::ZlibTxListCodec;
+use tracing::{debug, warn};
 
 use crate::{
-    codec::{DecodedUnsafePayload, WhitelistExecutionPayloadEnvelope},
-    error::Result,
+    codec::{
+        DecodedUnsafePayload, MAX_COMPRESSED_TX_LIST_BYTES, MAX_DECOMPRESSED_TX_LIST_BYTES,
+        WhitelistExecutionPayloadEnvelope,
+    },
+    error::{Result, WhitelistPreconfirmationDriverError},
     metrics::WhitelistPreconfirmationDriverMetrics,
+    network::NetworkCommand,
 };
 
 use super::{
@@ -112,20 +120,185 @@ where
         Ok(())
     }
 
-    /// Handle a block-hash request from the request topic.
-    pub(super) async fn handle_unsafe_request(&mut self, hash: B256) -> Result<()> {
-        let (envelope, result_label) = if let Some(envelope) = self.state.get_recent(&hash).await {
+    /// Serve an envelope by block hash from the recent cache or local L2 state,
+    /// publishing it on the response topic.
+    ///
+    /// Returns the source label of the served envelope, or `None` when the hash
+    /// is not servable. `mark_end_of_sequencing` tags envelopes rebuilt from L2
+    /// so EOS catch-up responses carry the marker.
+    pub(super) async fn serve_envelope_by_hash(
+        &mut self,
+        hash: B256,
+        mark_end_of_sequencing: bool,
+    ) -> Result<Option<&'static str>> {
+        let (envelope, source) = if let Some(envelope) = self.state.get_recent(&hash).await {
             (envelope, "cache_hit")
-        } else if let Some(envelope) = self.build_response_envelope_from_l2(hash).await? {
+        } else if let Some(mut envelope) = self.build_response_envelope_from_l2(hash).await? {
+            if mark_end_of_sequencing {
+                envelope.end_of_sequencing = Some(true);
+            }
             (Arc::new(envelope), "l2_hit")
         } else {
-            WhitelistPreconfirmationDriverMetrics::inc_response_lookup("not_found");
-            return Ok(());
+            return Ok(None);
         };
 
-        WhitelistPreconfirmationDriverMetrics::inc_response_lookup(result_label);
         self.state.insert_recent(envelope.clone()).await;
         self.publish_unsafe_response(envelope).await;
+        Ok(Some(source))
+    }
+
+    /// Handle a block-hash request from the request topic.
+    pub(super) async fn handle_unsafe_request(&mut self, hash: B256) -> Result<()> {
+        let label = self.serve_envelope_by_hash(hash, false).await?.unwrap_or("not_found");
+        WhitelistPreconfirmationDriverMetrics::inc_response_lookup(label);
         Ok(())
+    }
+
+    /// Build a response envelope from local L2 state for an unsafe request hash.
+    async fn build_response_envelope_from_l2(
+        &self,
+        hash: B256,
+    ) -> Result<Option<WhitelistExecutionPayloadEnvelope>> {
+        let Some(block) = self
+            .rpc
+            .l2_provider
+            .get_block_by_hash(hash)
+            .full()
+            .await
+            .map_err(WhitelistPreconfirmationDriverError::provider)?
+            .map(|block| block.map_transactions(|tx: RpcTransaction| tx.into()))
+        else {
+            return Ok(None);
+        };
+
+        if self
+            .head_l1_origin_block_id()
+            .await?
+            .is_some_and(|head_l1_origin_block_id| block.header.number <= head_l1_origin_block_id)
+        {
+            return Ok(None);
+        }
+
+        let Some(l1_origin) = self.rpc.l1_origin_by_id(U256::from(block.header.number)).await?
+        else {
+            return Ok(None);
+        };
+
+        if l1_origin.signature == [0u8; 65] {
+            return Ok(None);
+        }
+
+        let Some(transactions) = block.transactions.as_transactions() else {
+            return Err(WhitelistPreconfirmationDriverError::invalid_payload(
+                "request-response block missing full transaction bodies",
+            ));
+        };
+
+        let raw_transactions = transactions
+            .iter()
+            .map(|tx: &TxEnvelope| tx.encoded_2718().to_vec())
+            .collect::<Vec<_>>();
+        let compressed_tx_list = ZlibTxListCodec::new_with_limits(
+            MAX_COMPRESSED_TX_LIST_BYTES,
+            MAX_DECOMPRESSED_TX_LIST_BYTES,
+        )
+        .encode(&raw_transactions)
+        .map_err(|err| {
+            WhitelistPreconfirmationDriverError::invalid_payload_with_context(
+                "failed to encode request-response tx list",
+                err,
+            )
+        })?;
+
+        let end_of_sequencing = self
+            .cache
+            .get(&hash)
+            .and_then(|envelope| envelope.end_of_sequencing)
+            .filter(|&enabled| enabled);
+        let base_fee = block.header.base_fee_per_gas.ok_or_else(|| {
+            WhitelistPreconfirmationDriverError::invalid_payload(format!(
+                "request-response block {} missing base fee",
+                block.header.number
+            ))
+        })?;
+
+        Ok(Some(WhitelistExecutionPayloadEnvelope {
+            end_of_sequencing,
+            is_forced_inclusion: l1_origin.is_forced_inclusion.then_some(true),
+            // Intentionally None — the sequencer never populates this field
+            // in the envelope (see rest_handler.rs), and the SSZ wire format
+            // encodes None as 32 zero bytes which is the expected default.
+            parent_beacon_block_root: None,
+            // Carry Unzen header.difficulty (= block_zk_gas_used) so receivers
+            // can reconstruct the sender's block hash. Left `None` for Shasta
+            // blocks whose difficulty is zero.
+            header_difficulty: (!block.header.difficulty.is_zero())
+                .then_some(block.header.difficulty),
+            execution_payload: alloy_rpc_types_engine::ExecutionPayloadV1 {
+                parent_hash: block.header.parent_hash,
+                fee_recipient: block.header.beneficiary,
+                state_root: block.header.state_root,
+                receipts_root: block.header.receipts_root,
+                logs_bloom: block.header.logs_bloom,
+                prev_randao: block.header.mix_hash,
+                block_number: block.header.number,
+                gas_limit: block.header.gas_limit,
+                gas_used: block.header.gas_used,
+                timestamp: block.header.timestamp,
+                extra_data: block.header.extra_data.clone(),
+                base_fee_per_gas: U256::from(base_fee),
+                block_hash: block.header.hash,
+                transactions: vec![Bytes::from(compressed_tx_list)],
+            },
+            // Use L1-origin signature as-is. This endpoint serves responses
+            // only from the local node; caller-side validation and allowlist
+            // checks are performed when importing the envelope.
+            signature: Some(l1_origin.signature),
+        }))
+    }
+
+    /// Publish a block-hash request on `requestPreconfBlocks`.
+    pub(super) async fn publish_unsafe_request(&self, hash: B256) {
+        if let Err(err) =
+            self.network_command_tx.send(NetworkCommand::PublishUnsafeRequest { hash }).await
+        {
+            WhitelistPreconfirmationDriverMetrics::inc_network_outbound_publish(
+                "request_preconf_blocks",
+                "queue_failed",
+            );
+            warn!(
+                hash = %hash,
+                error = %err,
+                "failed to queue whitelist preconfirmation request publish command"
+            );
+        } else {
+            WhitelistPreconfirmationDriverMetrics::inc_network_outbound_publish(
+                "request_preconf_blocks",
+                "queued",
+            );
+        }
+    }
+
+    /// Publish an envelope response on `responsePreconfBlocks`.
+    async fn publish_unsafe_response(&self, envelope: Arc<WhitelistExecutionPayloadEnvelope>) {
+        let hash = envelope.execution_payload.block_hash;
+        if let Err(err) =
+            self.network_command_tx.send(NetworkCommand::PublishUnsafeResponse { envelope }).await
+        {
+            WhitelistPreconfirmationDriverMetrics::inc_network_outbound_publish(
+                "response_preconf_blocks",
+                "queue_failed",
+            );
+            warn!(
+                hash = %hash,
+                error = %err,
+                "failed to queue whitelist preconfirmation response publish command"
+            );
+        } else {
+            WhitelistPreconfirmationDriverMetrics::inc_network_outbound_publish(
+                "response_preconf_blocks",
+                "queued",
+            );
+        }
     }
 }
