+++ packages/taiko-client-rs/crates/whitelist-preconfirmation-driver/src/network/runtime.rs
@@ -807,7 +807,11 @@ impl NetworkRuntime {
                         self.inbound_validation_state.validate_preconf_blocks(&payload);
 
                     if matches!(acceptance, gossipsub::MessageAcceptance::Accept) {
-                        log_inbound_preconf_blocks_entry(from, &payload);
+                        log_inbound_envelope(
+                            from,
+                            &payload.envelope,
+                            "📥 New preconfirmation block gossip",
+                        );
                         if let Err(err) = forward_event(
                             &self.event_tx,
                             NetworkEvent::UnsafePayload { from, payload },
@@ -840,7 +844,11 @@ impl NetworkRuntime {
                 Ok(envelope) => {
                     let acceptance = self.inbound_validation_state.validate_response(&envelope);
                     if matches!(acceptance, gossipsub::MessageAcceptance::Accept) {
-                        log_inbound_preconf_response_entry(from, &envelope);
+                        log_inbound_envelope(
+                            from,
+                            &envelope,
+                            "📥 New preconfirmation block response gossip",
+                        );
                         if let Err(err) = forward_event(
                             &self.event_tx,
                             NetworkEvent::UnsafeResponse { from, envelope },
@@ -878,7 +886,7 @@ impl NetworkRuntime {
                 return Ok(());
             };
 
-            let acceptance = self.inbound_validation_state.validate_request(from, hash, now);
+            let acceptance = self.inbound_validation_state.validate_request(from, now);
             if matches!(acceptance, gossipsub::MessageAcceptance::Accept) {
                 info!(
                     peer = %from,
@@ -1125,28 +1133,12 @@ fn acceptance_label(acceptance: &gossipsub::MessageAcceptance) -> &'static str {
     }
 }
 
-/// Emit an entry log for an inbound `preconfBlocks` gossip payload.
-fn log_inbound_preconf_blocks_entry(from: PeerId, payload: &DecodedUnsafePayload) {
-    let execution_payload = &payload.envelope.execution_payload;
-    info!(
-        peer = %from,
-        block_id = execution_payload.block_number,
-        block_hash = %execution_payload.block_hash,
-        coinbase = %execution_payload.fee_recipient,
-        timestamp = execution_payload.timestamp,
-        gas_limit = execution_payload.gas_limit,
-        gas_used = execution_payload.gas_used,
-        base_fee_per_gas = %execution_payload.base_fee_per_gas,
-        extra_data = %alloy_primitives::hex::encode(&execution_payload.extra_data),
-        parent_hash = %execution_payload.parent_hash,
-        end_of_sequencing = payload.envelope.end_of_sequencing.unwrap_or(false),
-        is_forced_inclusion = payload.envelope.is_forced_inclusion.unwrap_or(false),
-        "📥 New preconfirmation block gossip"
-    );
-}
-
-/// Emit an entry log for an inbound `responsePreconfBlocks` gossip payload.
-fn log_inbound_preconf_response_entry(from: PeerId, envelope: &WhitelistExecutionPayloadEnvelope) {
+/// Emit an entry log for an inbound gossip envelope.
+fn log_inbound_envelope(
+    from: PeerId,
+    envelope: &WhitelistExecutionPayloadEnvelope,
+    message: &'static str,
+) {
     let execution_payload = &envelope.execution_payload;
     info!(
         peer = %from,
@@ -1161,7 +1153,7 @@ fn log_inbound_preconf_response_entry(from: PeerId, envelope: &WhitelistExecutio
         parent_hash = %execution_payload.parent_hash,
         end_of_sequencing = envelope.end_of_sequencing.unwrap_or(false),
         is_forced_inclusion = envelope.is_forced_inclusion.unwrap_or(false),
-        "📥 New preconfirmation block response gossip"
+        "{message}"
     );
 }
